chore: refactor review sidebar
